fix : clean dividing file, moving checking built in func in execve part
